+++ app/core/routers.py
@@ -1,6 +1,7 @@
 from fastapi import APIRouter
 
-from app.modules.wage.routers import admin_router
+from app.modules.admin.routers import users_router, admin_router
+from app.modules.wage.routers import admin_router as wage_admin_router
 
 api_router = APIRouter()
 
@@ -13,11 +14,31 @@ routers = [
     ("/workstatus", "근태관리", "app.modules.workstatus.routers"),
     ("/community", "커뮤니티관리", "app.modules.community.routers"),
     ("/admin", "관리자", "app.modules.admin.routers"),
-    ("/wage", "유저시급", "app.modules.wage.routers"),
+    ("/wage", "유저관리", "app.modules.wage.routers"),
 ]
 
 for prefix, tag, module_path in routers:
     module = __import__(module_path, fromlist=["router"])
     api_router.include_router(module.router, prefix=prefix, tags=[tag])
 
-api_router.include_router(admin_router, prefix="/admin/default-wage", tags=["Admin"])
+
+# 관리자 - 유저관리
+api_router.include_router(
+    users_router,
+    prefix="/admin",
+    tags=["유저관리"],
+)
+
+# 관리자 - 관리자 기능
+api_router.include_router(
+    admin_router,
+    prefix="/admin",
+    tags=["관리자"],
+)
+
+# 관리자 - 최저시급 (기존 유지)
+api_router.include_router(
+    wage_admin_router,
+    prefix="/admin/default-wage",
+    tags=["최저시급 관리"],
+)
