+++ docker/docker-compose.dev.yml
@@ -11,7 +11,7 @@ services:
     environment:
       - TZ=Asia/Seoul
     volumes:
-      - ..:/app
+      - ../app:/app/app
     depends_on:
       - db
     networks:
